requirements.txt
tqdm>=4.51.0
paddlepaddle_gpu>=2.1.2
numpy>=1.20.3